done some errors

+++ src/main/java/studio/demo/service/impl/CommentServiceImpl.java
@@ -3,6 +3,9 @@ package studio.demo.service.impl;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
+import studio.demo.exception.CommentNullException;
+import studio.demo.exception.CommentWithThisNameDoesNotExist;
+import studio.demo.exception.CommentWithThisTopicExist;
 import studio.demo.model.entity.Comment;
 import studio.demo.model.service.CommentServiceModel;
 import studio.demo.model.view.CommentViewModel;
@@ -23,9 +26,22 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public void addComment(CommentServiceModel commentServiceModel) {
+    public CommentServiceModel addComment(CommentServiceModel commentServiceModel) throws CommentNullException, CommentWithThisTopicExist {
+
+        if (commentServiceModel.getTopic().trim().isEmpty()) {
+            throw new CommentNullException("Comment is null");
+        }
+
+        if (commentRepository.findById(commentServiceModel.getTopic()).isPresent()) {
+            throw new CommentWithThisTopicExist("Comment with this topic exists!");
+        }
+
         Comment comment = this.modelMapper.map(commentServiceModel, Comment.class);
+        comment.setTopic(commentServiceModel.getTopic());
+        comment.setDescription(commentServiceModel.getDescription());
         this.commentRepository.saveAndFlush(comment);
+
+        return commentServiceModel;
     }
 
     @Override
@@ -53,7 +69,36 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public void delete(String id) {
+    public boolean delete(String id) {
+        if (this.commentRepository.findById(id).isEmpty()) {
+            return false;
+        }
         this.commentRepository.deleteById(id);
+        return true;
     }
+
+    @Override
+    public CommentServiceModel update(CommentViewModel comment) throws CommentWithThisNameDoesNotExist {
+
+        Comment c = this.commentRepository.
+                findById(comment.getId()).orElse(null);
+        this.checkCommentExist(c);
+        if (!comment.getDescription().trim().isEmpty()) {
+            c.setDescription(comment.getDescription());
+        }
+        if (!comment.getTopic().trim().isEmpty()) {
+            c.setTopic(comment.getTopic());
+        }
+
+        return this.modelMapper.map(this.commentRepository.saveAndFlush(c),
+                CommentServiceModel.class);
+
+    }
+
+    private void checkCommentExist(Comment c) throws CommentWithThisNameDoesNotExist {
+        if (c == null) {
+            throw new CommentWithThisNameDoesNotExist("Comment with this name does not exist!");
+        }
+    }
+
 }
